# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CreatorOS API
  version: 0.10.0
  description: REST contract for the CreatorOS modular monolith.
servers:
  - url: http://localhost:8080
    description: Local API
security: []
...
paths:
...
  /api/v1/orders:
    get:
      operationId: listClientOrders
      summary: List authenticated client's orders
      tags: [Orders]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: List of client orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createOrder
      summary: Place an order for a published creator service package
      tags: [Orders]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order placed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InvalidCSRF"
        "404":
          $ref: "#/components/responses/ServiceNotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/orders/{order_id}:
    get:
      operationId: getClientOrder
      summary: Get order detail with brief versions and events as client
      tags: [Orders]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
      responses:
        "200":
          description: Order detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/OrderNotFound"
  /api/v1/orders/{order_id}/cancel:
    post:
      operationId: cancelOrder
      summary: Cancel an order before it enters progress
      tags: [Orders]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderActionRequest"
      responses:
        "200":
          description: Order cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InvalidCSRF"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          $ref: "#/components/responses/InvalidOrderTransition"
  /api/v1/orders/{order_id}/briefs:
    post:
      operationId: submitOrderBrief
      summary: Submit a new versioned brief for an order
      tags: [Orders]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitBriefRequest"
      responses:
        "201":
          description: New brief version submitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BriefVersionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InvalidCSRF"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          $ref: "#/components/responses/InvalidOrderTransition"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/creator/orders:
    get:
      operationId: listCreatorOrders
      summary: List incoming orders for authenticated creator
      tags: [Orders]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: List of creator orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/creator/orders/{order_id}:
    get:
      operationId: getCreatorOrder
      summary: Get order detail with brief versions and events as creator
      tags: [Orders]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
      responses:
        "200":
          description: Creator order detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/OrderNotFound"
  /api/v1/creator/orders/{order_id}/accept:
    post:
      operationId: acceptOrder
      summary: Accept an incoming order and establish project deadline
      tags: [Orders]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderActionRequest"
      responses:
        "200":
          description: Order accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InvalidCSRF"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          $ref: "#/components/responses/InvalidOrderTransition"
  /api/v1/creator/orders/{order_id}/decline:
    post:
      operationId: declineOrder
      summary: Decline an incoming order
      tags: [Orders]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderActionRequest"
      responses:
        "200":
          description: Order declined.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/InvalidCSRF"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          $ref: "#/components/responses/InvalidOrderTransition"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: creatoros_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Must match the creatoros_csrf cookie bound to the current session.
      schema:
        type: string
...
    OrderID:
      name: order_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The JSON body is malformed, too large, contains unknown fields, or is not a single object.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: One or more request fields are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: The account cannot perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthenticated:
      description: A valid active session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidCSRF:
      description: The CSRF cookie and header are missing, different, or not bound to the session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceNotFound:
      description: The service does not exist, is not owned by the caller, or is not public.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    OrderNotFound:
      description: The order was not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    InvalidOrderTransition:
      description: The order cannot transition from its current state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    OrderStatus:
      type: string
      enum:
        - pending_acceptance
        - accepted
        - declined
        - in_progress
        - completed
        - cancelled
        - disputed
    Order:
      type: object
      required:
        - id
        - client_user_id
        - creator_user_id
        - service_id
        - package_id
        - package_name
        - package_description
        - price_minor
        - currency
        - delivery_days
        - revision_limit
        - brief_content
        - status
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        client_user_id: { type: string, format: uuid }
        client_display_name: { type: string }
        creator_user_id: { type: string, format: uuid }
        creator_display_name: { type: string }
        creator_slug: { type: string }
        service_id: { type: string, format: uuid }
        service_title: { type: string }
        package_id: { type: string, format: uuid }
        package_name: { type: string }
        package_description: { type: string }
        price_minor: { type: integer, format: int64, minimum: 1 }
        currency: { type: string, enum: [IDR, MYR, USD] }
        delivery_days: { type: integer, minimum: 1, maximum: 365 }
        revision_limit: { type: integer, minimum: 0, maximum: 20 }
        brief_content: { type: string }
        status: { $ref: "#/components/schemas/OrderStatus" }
        accepted_at:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        deadline_at:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        cancelled_at:
          oneOf: [{ type: string, format: date-time }, { type: "null" }]
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BriefVersion:
      type: object
      required:
        - id
        - order_id
        - version
        - content
        - submitted_by
        - created_at
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        content: { type: string }
        submitted_by: { type: string, format: uuid }
        submitted_by_name: { type: string }
        created_at: { type: string, format: date-time }
    OrderEvent:
      type: object
      required:
        - id
        - order_id
        - actor_user_id
        - from_status
        - to_status
        - created_at
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        actor_user_id: { type: string, format: uuid }
        actor_name: { type: string }
        from_status: { type: string }
        to_status: { type: string }
        note: { type: string }
        created_at: { type: string, format: date-time }
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          required: [briefs, events]
          properties:
            briefs:
              type: array
              items: { $ref: "#/components/schemas/BriefVersion" }
            events:
              type: array
              items: { $ref: "#/components/schemas/OrderEvent" }
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [service_id, package_id, brief_content]
      properties:
        service_id: { type: string, format: uuid }
        package_id: { type: string, format: uuid }
        brief_content: { type: string, minLength: 20, maxLength: 5000 }
    OrderActionRequest:
      type: object
      additionalProperties: false
      properties:
        note: { type: string, maxLength: 1000 }
    SubmitBriefRequest:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content: { type: string, minLength: 20, maxLength: 5000 }
    OrderResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/Order" }
    OrderDetailResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/OrderDetail" }
    OrderListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Order" }
    BriefVersionResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/BriefVersion" }
...